pyproject.toml
[project]
name = "llm-analysis-quiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.38.0",
    "webdriver-manager>=4.0.2",
]
